+++ client/views/Issue/IssueCardDetailBody.jsx
@@ -68,17 +68,17 @@ IssueDynamicArea = React.createClass({
                     <IpsoTabso tabs={[
                         {
                             tab:<span><span>Files</span>{this.data.attachmentCount?<span>({this.data.attachmentCount})</span>:null}</span>,
-                            content:<div style={{padding:"20px"}}>
+                            content:<div style={{padding:"15px"}}>
                                 <AutoForm item={issue} schema={Issues.schema()} form={['attachments']} save={this.props.save} />
                             </div>
                         },{
                             tab:<span><span>Contacts</span></span>,
-                            content:<div style={{padding:"20px"}}>
+                            content:<div style={{padding:"15px"}}>
                                 <ContactList items={contacts}/>
                             </div>
                         },{
                             tab:<span><span>Updates</span>{this.data.messageCount?<span>({this.data.messageCount})</span>:null}</span>,
-                            content:<div style={{padding:"20px"}}>
+                            content:<div style={{padding:"15px"}}>
                                 <Inbox for={issue} />
                             </div>
                         }